+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to
 * Changed `EthernetServer::end()` to return `void` instead of `bool`.
 * Changed `MDNSClass::begin(hostname)` and `DNSClient::getHostByName()` to treat
   a NULL hostname as an error; they now explicitly return false in this case.
+* Changed `MDNSClass::end()` to return `void` instead of `bool`.
 
 ### Removed
 * `EthernetServer` and `EthernetUDP` begin functions that take a Boolean